+++ packages/bridgette-twitter/index.js
@@ -1,6 +1,8 @@
-require("dotenv").config();
 var { log, T } =require("@chippr-bots/common");
 var bdb = require("@chippr-bots/bridgettedb");
+const path = require('path')
+
+var {training, character} = require("./mode");
 
 var {
   tags,
@@ -8,7 +10,6 @@ var {
   } = require("./lib");
 
 log.info("🤖  bridgette-twitter loaded with DBKEY: " + process.env.BDB_DBKEY);
-
 
 var b = new bdb({
   "nodeAddr": process.env.BDB_NODE_URL,
@@ -18,26 +19,48 @@ var b = new bdb({
   "DBKEY": process.env.BDB_DBKEY
   })
 
-var local = {
-  local.likeTH : local.likeTH,
-  local.likeTH
+var state = {
+  "activeState" : "training",
+  tweetstack : [],
+  likeTH : 9999999,
+  rtTH : 9999999,
+  hashtags: [],
+  naughty: []
+}
+
+
+//keep refreshing lists
+setInterval(() => {
+  try {
+    b.get("likeTH").then(res => {log.debug(`o likeTH: ${res}`); state.likeTH = parseInt(res,10)});
+    b.get("rtTH").then(res => {log.debug(`o rtTH: ${res}` ); state.rtTH = parseInt(res, 10)});
+    b.get("nice").then(res => {log.debug(`o Nice: ${res}` ); state.hashtags = res.split(",")});
+    b.get("naughty").then(res => {log.debug(`o Naughty: ${res}` ); state.naughty = res.split(",")});
+    b.get("activeState").then(res => {log.debug(`o State: ${res}` ); state.activeState = res});
+  } catch (error) {
+    log.error(`[bridgette-twitter/index] error getting variables: ${error}`)
+  }
+}, process.env.SYNC_LOOP_TIMER);
+
+function main(){
+  log.info(`o likeTH: ${state.likeTH} | rtTH: ${state.rtTH} | Nice ${state.hashtags} | Naughty ${state.naughty}` );
+  switch (state.activeState) {
+    case "training":
+        let tl = state.hashtags.concat(state.naughty);
+        log.info( `training on: ${tl}`);
+        training(T, tl);
+      break;
+    case "character":
+      //log mode 
+      log.info( `Tweeting on: ${state.hashtags}`);
+      //Run loop
+      character(T, state);      
+      break;
+    default:
+      break;
+    }
 }
 
 setInterval(() => {
-     b.get(process.env.DBKEY + "likeTH").then(res => {log.info("o likeTH: " + res ); b.likeTH = parseInt(res,10)});
-     b.get(process.env.DBKEY + "rtTH").then(res => {log.info("o rtTH: " + res ); b.rtTH = parseInt(res, 10)});
-     b.get(process.env.DBKEY + "nice").then(res => {log.info("o Nice: " + res ); b.hashtags = res.split(",")});
-     b.get(process.env.DBKEY + "naughty").then(res => {log.info("o Naughty: " + res ); b.naughty = res.split(",")});
-}, 6000);
-
-
-
-function main(){
-  for(tag in b.hashtags){
-    let res = seeker(T, b.hashtags[tag], b.naughty, b.likeTH, b.rtTH);
-   }
-}
-
-main();
-
-setInterval(() => {main()}, 600000);
+  main()
+}, process.env.MAIN_LOOP_TIMER);