+++ packages/bridgette-twitter/lib/index.js
@@ -1,6 +1,12 @@
 
 var seeker = require("./seeker");
+var bayes = require("./Bayes");
+var markov = require("./markov");
+var getModel = require("./getModel");
 
 module.exports = {
   seeker : seeker,
+  bayes : bayes,
+  markov : markov,
+  getModel : getModel
 }