+++ packages/bridgette-twitter/mode/index.js
@@ -1,0 +1,12 @@
+
+var character = require('./character');
+var training = require('./training');
+var sleep = require('./sleep');
+var debug = require('./debug');
+
+module.exports = {
+    character : character,
+    training : training,
+    sleep : sleep,
+    debug : debug
+}