+++ packages/bridgette-twitter/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.45](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.44...v1.1.45) (2020-05-05)
+
+**Note:** Version bump only for package @chippr-bots/bridgette-twitter
+
+
+
+
+
+## [1.1.44](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.43...v1.1.44) (2020-04-17)
+
+**Note:** Version bump only for package @chippr-bots/bridgette-twitter
+
+
+
+
+
+## [1.1.43](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.42...v1.1.43) (2020-04-17)
+
+**Note:** Version bump only for package @chippr-bots/bridgette-twitter
+
+
+
+
+
 ## [1.1.42](https://github.com/chippr-robotics/chippr-bots/compare/v1.1.41...v1.1.42) (2020-04-13)
 
 **Note:** Version bump only for package @chippr-bots/bridgette-twitter